Refactor logging routes into separate module for improved organization
- Moved logging-related API endpoints from `web.rs` to a new `logs.rs` module, enhancing code organization and maintainability.
- Updated routing in `web.rs` to integrate the new logging routes, ensuring functionality remains intact.
- Improved readability and separation of concerns by isolating logging logic from the main web handler.

These changes align with best practices for code structure and maintainability, facilitating easier future enhancements to the logging system.
